FZ10 pull request

+++ src/main/java/ru/java/practicum/filmorate/model/BaseUnit.java
@@ -5,10 +5,12 @@ import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
+
 @Data
 @EqualsAndHashCode(of = {"id"}) //для сравнения используем только поле id
 @SuperBuilder    //аннотация которая позволяет создавать классы билдер для классов наследников
 @NoArgsConstructor
 public abstract class BaseUnit {
-    private int id;
+    private long id;
+
 }
